Fix Basic Editing At End Of Files & New Files

+++ Sources/CodeEditInputView/TextLayoutManager/TextLayoutManager+Edits.swift
@@ -21,10 +21,10 @@ extension TextLayoutManager: NSTextStorageDelegate {
     ///   - string: The string to replace in the given range.
     public func willReplaceCharactersInRange(range: NSRange, with string: String) {
         // Loop through each line being replaced in reverse, updating and removing where necessary.
-        for linePosition in lineStorage.linesInRange(range).reversed() {
+         for linePosition in lineStorage.linesInRange(range).reversed() {
             // Two cases: Updated line, deleted line entirely
             guard let intersection = linePosition.range.intersection(range), !intersection.isEmpty else { continue }
-            if intersection == linePosition.range {
+            if intersection == linePosition.range && linePosition.range.max != lineStorage.length {
                 // Delete line
                 lineStorage.delete(lineAt: linePosition.range.location)
             } else if intersection.max == linePosition.range.max,
@@ -57,7 +57,6 @@ extension TextLayoutManager: NSTextStorageDelegate {
                 )
             }
         }
-
         setNeedsLayout()
     }
 
@@ -66,30 +65,34 @@ extension TextLayoutManager: NSTextStorageDelegate {
     ///   - insertedString: The string being inserted.
     ///   - location: The location the string is being inserted into.
     private func applyLineInsert(_ insertedString: NSString, at location: Int) {
-        if lineStorage.count == 0 && lineStorage.length == 0 {
-            // The text was completely empty before, insert.
-            lineStorage.insert(
-                line: TextLine(),
-                atIndex: location,
-                length: insertedString.length,
-                height: estimateLineHeight()
-            )
-        } else if LineEnding(line: insertedString as String) != nil {
-            // Need to split the line inserting into and create a new line with the split section of the line
-            guard let linePosition = lineStorage.getLine(atIndex: location) else { return }
-            let splitLocation = location + insertedString.length
-            let splitLength = linePosition.range.max - location
-            let lineDelta = insertedString.length - splitLength // The difference in the line being edited
-            if lineDelta != 0 {
-                lineStorage.update(atIndex: location, delta: lineDelta, deltaHeight: 0.0)
-            }
+        if LineEnding(line: insertedString as String) != nil {
+            if location == textStorage.length {
+                // Insert a new line at the end of the document, need to insert a new line 'cause there's nothing to
+                // split. Also, append the new text to the last line.
+                lineStorage.update(atIndex: location, delta: insertedString.length, deltaHeight: 0.0)
+                lineStorage.insert(
+                    line: TextLine(),
+                    asOffset: location + insertedString.length,
+                    length: 0,
+                    height: estimateLineHeight()
+                )
+            } else {
+                // Need to split the line inserting into and create a new line with the split section of the line
+                guard let linePosition = lineStorage.getLine(atIndex: location) else { return }
+                let splitLocation = location + insertedString.length
+                let splitLength = linePosition.range.max - location
+                let lineDelta = insertedString.length - splitLength // The difference in the line being edited
+                if lineDelta != 0 {
+                    lineStorage.update(atIndex: location, delta: lineDelta, deltaHeight: 0.0)
+                }
 
-            lineStorage.insert(
-                line: TextLine(),
-                atIndex: splitLocation,
-                length: splitLength,
-                height: estimateLineHeight()
-            )
+                lineStorage.insert(
+                    line: TextLine(),
+                    asOffset: splitLocation,
+                    length: splitLength,
+                    height: estimateLineHeight()
+                )
+            }
         } else {
             lineStorage.update(atIndex: location, delta: insertedString.length, deltaHeight: 0.0)
         }
@@ -108,4 +111,141 @@ extension TextLayoutManager: NSTextStorageDelegate {
             invalidateLayoutForRange(editedRange)
         }
     }
+}
+
+extension TextLineStorage {
+    var description: String {
+        treeString(root!) {
+            ("\($0.length)", $0.left, $0.right)
+        }
+    }
+}
+
+public func treeString<T>(_ node:T, reversed:Bool=false, isTop:Bool=true, using nodeInfo:(T)->(String,T?,T?)) -> String
+{
+    // node value string and sub nodes
+    let (stringValue, leftNode, rightNode) = nodeInfo(node)
+
+    let stringValueWidth  = stringValue.count
+
+    // recurse to sub nodes to obtain line blocks on left and right
+    let leftTextBlock     = leftNode  == nil ? []
+    : treeString(leftNode!,reversed:reversed,isTop:false,using:nodeInfo)
+        .components(separatedBy:"\n")
+
+    let rightTextBlock    = rightNode == nil ? []
+    : treeString(rightNode!,reversed:reversed,isTop:false,using:nodeInfo)
+        .components(separatedBy:"\n")
+
+    // count common and maximum number of sub node lines
+    let commonLines       = min(leftTextBlock.count,rightTextBlock.count)
+    let subLevelLines     = max(rightTextBlock.count,leftTextBlock.count)
+
+    // extend lines on shallower side to get same number of lines on both sides
+    let leftSubLines      = leftTextBlock
+    + Array(repeating:"", count: subLevelLines-leftTextBlock.count)
+    let rightSubLines     = rightTextBlock
+    + Array(repeating:"", count: subLevelLines-rightTextBlock.count)
+
+    // compute location of value or link bar for all left and right sub nodes
+    //   * left node's value ends at line's width
+    //   * right node's value starts after initial spaces
+    let leftLineWidths    = leftSubLines.map{$0.count}
+    let rightLineIndents  = rightSubLines.map{$0.prefix{$0==" "}.count  }
+
+    // top line value locations, will be used to determine position of current node & link bars
+    let firstLeftWidth    = leftLineWidths.first   ?? 0
+    let firstRightIndent  = rightLineIndents.first ?? 0
+
+
+    // width of sub node link under node value (i.e. with slashes if any)
+    // aims to center link bars under the value if value is wide enough
+    //
+    // ValueLine:    v     vv    vvvvvv   vvvvv
+    // LinkLine:    / \   /  \    /  \     / \
+    //
+    let linkSpacing       = min(stringValueWidth, 2 - stringValueWidth % 2)
+    let leftLinkBar       = leftNode  == nil ? 0 : 1
+    let rightLinkBar      = rightNode == nil ? 0 : 1
+    let minLinkWidth      = leftLinkBar + linkSpacing + rightLinkBar
+    let valueOffset       = (stringValueWidth - linkSpacing) / 2
+
+    // find optimal position for right side top node
+    //   * must allow room for link bars above and between left and right top nodes
+    //   * must not overlap lower level nodes on any given line (allow gap of minSpacing)
+    //   * can be offset to the left if lower subNodes of right node
+    //     have no overlap with subNodes of left node
+    let minSpacing        = 2
+    let rightNodePosition = zip(leftLineWidths,rightLineIndents[0..<commonLines])
+        .reduce(firstLeftWidth + minLinkWidth)
+    { max($0, $1.0 + minSpacing + firstRightIndent - $1.1) }
+
+
+    // extend basic link bars (slashes) with underlines to reach left and right
+    // top nodes.
+    //
+    //        vvvvv
+    //       __/ \__
+    //      L       R
+    //
+    let linkExtraWidth    = max(0, rightNodePosition - firstLeftWidth - minLinkWidth )
+    let rightLinkExtra    = linkExtraWidth / 2
+    let leftLinkExtra     = linkExtraWidth - rightLinkExtra
+
+    // build value line taking into account left indent and link bar extension (on left side)
+    let valueIndent       = max(0, firstLeftWidth + leftLinkExtra + leftLinkBar - valueOffset)
+    let valueLine         = String(repeating:" ", count:max(0,valueIndent))
+    + stringValue
+    let slash             = reversed ? "\\" : "/"
+    let backSlash         = reversed ? "/"  : "\\"
+    let uLine             = reversed ? "¯"  : "_"
+    // build left side of link line
+    let leftLink          = leftNode == nil ? ""
+    : String(repeating: " ", count:firstLeftWidth)
+    + String(repeating: uLine, count:leftLinkExtra)
+    + slash
+
+    // build right side of link line (includes blank spaces under top node value)
+    let rightLinkOffset   = linkSpacing + valueOffset * (1 - leftLinkBar)
+    let rightLink         = rightNode == nil ? ""
+    : String(repeating:  " ", count:rightLinkOffset)
+    + backSlash
+    + String(repeating:  uLine, count:rightLinkExtra)
+
+    // full link line (will be empty if there are no sub nodes)
+    let linkLine          = leftLink + rightLink
+
+    // will need to offset left side lines if right side sub nodes extend beyond left margin
+    // can happen if left subtree is shorter (in height) than right side subtree
+    let leftIndentWidth   = max(0,firstRightIndent - rightNodePosition)
+    let leftIndent        = String(repeating:" ", count:leftIndentWidth)
+    let indentedLeftLines = leftSubLines.map{ $0.isEmpty ? $0 : (leftIndent + $0) }
+
+    // compute distance between left and right sublines based on their value position
+    // can be negative if leading spaces need to be removed from right side
+    let mergeOffsets      = indentedLeftLines
+        .map{$0.count}
+        .map{leftIndentWidth + rightNodePosition - firstRightIndent - $0 }
+        .enumerated()
+        .map{ rightSubLines[$0].isEmpty ? 0  : $1 }
+
+
+    // combine left and right lines using computed offsets
+    //   * indented left sub lines
+    //   * spaces between left and right lines
+    //   * right sub line with extra leading blanks removed.
+    let mergedSubLines    = zip(mergeOffsets.enumerated(),indentedLeftLines)
+        .map{ ( $0.0, $0.1, $1 + String(repeating:" ", count:max(0,$0.1)) ) }
+        .map{ $2 + String(rightSubLines[$0].dropFirst(max(0,-$1))) }
+
+    // Assemble final result combining
+    //  * node value string
+    //  * link line (if any)
+    //  * merged lines from left and right sub trees (if any)
+    let treeLines = [leftIndent + valueLine]
+    + (linkLine.isEmpty ? [] : [leftIndent + linkLine])
+    + mergedSubLines
+
+    return (reversed && isTop ? treeLines.reversed(): treeLines)
+        .joined(separator:"\n")
 }
